-> fp16

+++ inference_gpu_crypten.py
@@ -9,7 +9,7 @@ crypten.init()
 # Load the model and tokenizer
 model_name = "meta-llama/Meta-Llama-3-8B"
 tokenizer = AutoTokenizer.from_pretrained(model_name)
-model = LlamaModel.from_pretrained(model_name).to("cuda")  # Use LlamaModel instead of AutoModelForCausalLM
+model = LlamaModel.from_pretrained(model_name).to("cuda").half()  # Use FP16 for model
 
 # Define a wrapper for the CrypTen model
 class CrypTenLlamaModel(cnn.Module):
@@ -25,11 +25,11 @@ class CrypTenLlamaModel(cnn.Module):
 crypten_model = CrypTenLlamaModel(model).encrypt()
 
 # Load the decoder separately and keep it unencrypted
-decoder = AutoModelForCausalLM.from_pretrained(model_name).to("cuda")
+decoder = AutoModelForCausalLM.from_pretrained(model_name).to("cuda").half()  # Use FP16 for decoder
 
 # Prepare the input data
 input_text = "This is a test input."
-input_ids = tokenizer.encode(input_text, return_tensors="pt").to("cuda")
+input_ids = tokenizer.encode(input_text, return_tensors="pt").to("cuda").half()  # Use FP16 for input_ids
 input_ids_enc = crypten.cryptensor(input_ids)
 
 # Function to measure inference time with CrypTen (partial encryption)
@@ -42,7 +42,7 @@ def inference_crypten(crypten_model, decoder, input_ids_enc):
         # Forward pass through the encrypted model part
         hidden_states_enc = crypten_model(input_ids_enc)
         # Decrypt hidden states for the decoder
-        hidden_states = hidden_states_enc.get_plain_text()
+        hidden_states = hidden_states_enc.get_plain_text().half()  # Use FP16 for hidden states
         # Forward pass through the decoder
         outputs = decoder(inputs_embeds=hidden_states).logits
     torch.cuda.synchronize()  # Synchronize GPU after finishing the timer
